analyzer/requirements.txt
pyyaml
opencv-python
matplotlib
numpy
networkx